+++ datastore/dynamodb/extensions/datastores/dynamodb_datastore/lock.ts
@@ -7,7 +7,7 @@ import {
   GetCommand,
   PutCommand,
   UpdateCommand,
-} from "npm:@aws-sdk/lib-dynamodb@3.1121.0";
+} from "npm:@aws-sdk/lib-dynamodb@3.1126.0";
 import { lockKey } from "./keys.ts";
 import {
   Attr,